fix Day.js query error
